requirements.txt
requests==2.24.0
pytest==6.0.1
keyring==21.4.0
imapclient==2.1.0

selenium
pytest_selenium

locustio
invokust

paramiko
radar
forgery_py